+++ index.html
@@ -17,18 +17,15 @@
     />
     <title>Embeddable contact card widget</title>
 </head>
-<body>
+<body class="max-width center-screen">
 <div class="container contact-form-container"></div>
-<div id="contact-widget" class="box-shadow contact-widget"></div>
+<div id="contact-widget" class="contact-widget-show-container"></div>
 
 <!--<div class="flex col" id="root"></div>-->
 <!-- Phone Number DropDown & Validation -->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.8/js/intlTelInput.min.js"></script>
-<script src="src/scripts/contactCard.js"></script>
 
-<script src="index.js" type="module">
-  main().then(r => r).catch(r => r).finally(r => r);
-</script>
+<script src="index.js" type="module"></script>
 
 </body>
 </html>
