+++ index.js
@@ -1,11 +1,21 @@
-import {init} from "./src/scripts/init";
-import {loadContactCardWidget} from "./src/scripts/contactCard";
+import {attachEvents} from "./src/scripts/init.mjs";
+import {loadContactCardWidget} from "./src/scripts/contactCard.mjs";
+
+const promiseClosure = (e) => {
+  if (!e) {
+    return;
+  }
+  console.log("promise", e);
+};
 
 const main = async () => {
   const loadWidgetPromise = loadContactCardWidget();
   loadWidgetPromise.finally(() => {
     console.log("widget has been loaded");
-    init();
+    attachEvents();
   });
 };
-export default main;
+//invoke immediately
+(function () {
+  main().then(promiseClosure).catch(promiseClosure).finally(promiseClosure);
+})();
