+++ src/styles/contact-card.css
@@ -10,7 +10,7 @@
 .contact-widget {
     flex: 1;
     padding: var(--20px);
-    background: var(--color-white);
+    background-color: var(--color-white);
     border: 1px solid var(--color-grey);
     border-radius: var(--8px);
 }
